Revert the fix of bug 1438747
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@2965 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/Orts.Formats.Msts/SFile.cs
@@ -1397,14 +1397,14 @@ namespace Orts.Formats.Msts
     public class animation
     {
         public int FrameCount;          // :uint,num_frames
-        public float FrameRate;           // :uint,frame_rate 
+        public int FrameRate;           // :uint,frame_rate 
         public anim_nodes anim_nodes;    // :anim_nodes,AnimNodes .
 
         public animation(SBR block)
         {
             block.VerifyID(TokenID.animation);
             FrameCount = block.ReadInt();
-            FrameRate = block.ReadFloat();
+            FrameRate = block.ReadInt();
             anim_nodes = new anim_nodes(block.ReadSubBlock());
             block.VerifyEndOfBlock();
         }
